requirements.txt
spacy==2.2.3
newsapi-python==0.2.6
editdistance==0.5.3
snapy==1.0.2
mmh3==2.5.1
